Fix bugs in kettle_asset.xml to ensure compatibility with the latest MuJoCo versions
There's a missing <default> tag in kettle_asset.xml. This will bring errors in the latest version of MuJoCo (>3.16). I have validated that this change does not change the behavior of the environment for older versions of mujoco (3.1.6, 3.0.1). However, the behavior in MuJoCo version (<3.0, e.g. 2.1.5) remains to be checked.