Fix rgba value in oven_asset.xml to be in [0,1]
The mujoco xml docs state that "All [rgba] components should be in the
range [0 1]."
https://mujoco.readthedocs.io/en/stable/XMLreference.html#asset-material

MuJoCo's parser clearly let's this through. But it seems that mine is
a little more strict.

I checked with the original author @abhishekunique, and he confirmed
that it's just a bug (from his original work), and that it should be
fixed.